alx 3

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,12 +11,12 @@
 int main(void)
 {
 	int count;
-	int fib1 = 0, fib2 = 1, sum;
+	unsigned long int fib1 = 0, fib2 = 1, sum;
 
 	for (count = 0 ; count < 49 ; count++)
 	{
 		sum = fib1 + fib2;
-		printf("%d", sum);
+		printf("%lu", sum);
 
 		fib1 = fib2;
 		fib2 = sum;
